Delete product in store
- add test for bsdService.deleteProductsFromStore(1L, products);

	-- ProductsPriceInStore not deleted without exception (either store or
product relationships cancel deletion)
	-- remove ProductsPriceInStore from store as well and removal start
working

- add finders for product and store to bsdService

	- public Product findProductBySku(String string);
	- public Store findStoreById(Long l);

- add test deleteProductInStore() to validate bsd tables in suitcase

+++ MockExtJs6_bsd/MockServiceExtJs6_bsd/src/main/java/org/avp/bsd/service/BsdServiceImpl.java
@@ -3,6 +3,7 @@ package org.avp.bsd.service;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.Set;
 
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
@@ -156,6 +157,14 @@ public class BsdServiceImpl implements BsdService {
 		return ppis;
 	}
 
+	public Product findProductBySku(String sku){
+		return productRepository.findOne(sku);
+	}
+	public Store findStoreById(Long id){
+		return storeHeaderRepository.findOne(id);
+	}
+
+	
 	private Specification<ProductPriceInStore> findProductInStore( final Long storeId, final String sku) {
 		return new Specification<ProductPriceInStore>() {
 			@Override
@@ -194,13 +203,20 @@ public class BsdServiceImpl implements BsdService {
 	
 	@Transactional()
 	public void deleteProductsFromStore(Long storeId, List<ProductDto> products) throws Exception{
-		Store store = storeHeaderRepository.findOne(storeId);
+		/*Store store = storeHeaderRepository.findOne(storeId);
 		if (store == null) {
 			throw new Exception("Could not find '"+storeId+"' store id");
-		}
+		}*/
 		for (ProductDto productDto : products) {
 			ProductPriceInStore productPriceInStore = findProductPriceInStoreByStoreIdAndProductSku(storeId, productDto.getSku());
+			if(productPriceInStore == null){
+				continue;
+			}
+			Store store = productPriceInStore.getPk().getStore();
+			Set<ProductPriceInStore> productsInStore = store.getProductsInStore();
+			productsInStore.remove(productPriceInStore);
 			productPriceInStoreRepository.delete(productPriceInStore);
+			storeHeaderRepository.save(store);
 		}
 	}
